chore: refactor metrics
