+++ packages/bundler-relayer/src/metrics/index.ts
@@ -1,3 +1,2 @@
-export { MetricsHttpServer } from './metrics-server.js'
-export { MetricsTracker } from './metrics-tracker.js'
+export * from './metrics-tracker.js'
 export * from './metrics.types.js'
