+++ packages/bundler-relayer/src/metrics/influxdb/influxdb-client.ts
@@ -3,37 +3,43 @@ import os from 'os'
 import { InfluxDB, Point } from '@influxdata/influxdb-client'
 import { MeasurementName } from '../metrics.types.js'
 
-export class InfluxdbClient {
-  private readonly org: string
-  private readonly bucket: string
-  private readonly influxDB: InfluxDB
-
-  constructor(url: string, token: string, org: string, bucket: string) {
-    this.bucket = bucket
-    this.org = org
-    this.influxDB = new InfluxDB({ url, token })
-  }
-
-  public async writePoint(
+export type InfluxdbClient = {
+  writePoint: (
     measurement: string,
     fields: { value: number; name: MeasurementName }[],
-  ): Promise<void> {
-    const writeApi = this.influxDB.getWriteApi(this.org, this.bucket)
-    writeApi.useDefaultTags({ region: 'west' })
-    const timestamp = new Date().getTime() * 1000000 // In nanoseconds!
+  ) => Promise<void>
+}
 
-    const point = new Point(measurement)
-      .tag('host', os.hostname())
-      .tag('node_id', 'transeptor')
-      .timestamp(timestamp) // In nanoseconds!
+export const createInfluxdbClient = (
+  url: string,
+  token: string,
+  org: string,
+  bucket: string,
+) => {
+  const influxDB: InfluxDB = new InfluxDB({ url, token })
 
-    fields.forEach((field) => {
-      point.floatField(field.name, field.value)
-    })
+  return {
+    writePoint: async (
+      measurement: string,
+      fields: { value: number; name: MeasurementName }[],
+    ): Promise<void> => {
+      const writeApi = influxDB.getWriteApi(org, bucket)
+      writeApi.useDefaultTags({ region: 'west' })
+      const timestamp = new Date().getTime() * 1000000 // In nanoseconds!
 
-    writeApi.writePoint(point)
+      const point = new Point(measurement)
+        .tag('host', os.hostname())
+        .tag('node_id', 'transeptor')
+        .timestamp(timestamp) // In nanoseconds!
 
-    // Flush pending writes and close writeApi.
-    await writeApi.close()
+      fields.forEach((field) => {
+        point.floatField(field.name, field.value)
+      })
+
+      writeApi.writePoint(point)
+
+      // Flush pending writes and close writeApi.
+      await writeApi.close()
+    },
   }
 }
