+++ packages/shared/logger/base-logger.ts
@@ -2,35 +2,21 @@ import dotenv from 'dotenv'
 import pino, { LoggerOptions } from 'pino'
 dotenv.config()
 
-class Logger {
-  private static instance: pino.Logger
-
-  private constructor() {
-    //
+const createLogger = () => {
+  const options: LoggerOptions = {
+    level: process.env.NODE_ENV === 'production' ? 'info' : 'debug',
+    formatters: {
+      level: (label) => {
+        return { level: label.toUpperCase() }
+      },
+    },
+    redact: [
+      // Specify the field(s) to redact(client's IP address information will be redacted from the log output)
+      'req.headers.authorization',
+      'req.remoteAddress',
+    ],
   }
 
-  public static getInstance(): pino.Logger {
-    if (!Logger.instance) {
-      const options: LoggerOptions = {
-        level: process.env.NODE_ENV === 'production' ? 'info' : 'debug',
-        formatters: {
-          level: (label) => {
-            return { level: label.toUpperCase() }
-          },
-        },
-        redact: [
-          // Specify the field(s) to redact(client's IP address information will be redacted from the log output)
-          'req.headers.authorization',
-          'req.remoteAddress',
-        ],
-      }
-
-      Logger.instance = pino(options)
-    }
-
-    return Logger.instance
-  }
+  return pino(options)
 }
-
-const loggerInstance = Logger.getInstance()
-export { loggerInstance as Logger }
+export const Logger = createLogger()
